move files to different folders (#6)
* move files to different folders

* replace several usages of Vector3 with Bead and Component domain classes
