--- Assets/Scripts/Bead.cs
@@ -1,11 +1,0 @@
-using UnityEngine;
-
-public class Bead
-{
-    public Vector3 position;
-
-    public Bead(Vector3 position)
-    {
-        this.position = position;
-    }
-}
