+++ Assets/Scripts/BeadsProviders.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: df015a140a604ef1bc0a66fcf16591c4
+timeCreated: 1615053896
