+++ Assets/Scripts/BeadsProviders/ILinkBeadsProvider.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+using Domain;
+
+namespace BeadsProviders
+{
+    public interface ILinkBeadsProvider
+    {
+        List<LinkComponent> GetLinkComponents();
+    }
+}
