+++ Assets/Scripts/Domain/Bead.cs
@@ -1,0 +1,14 @@
+using UnityEngine;
+
+namespace Domain
+{
+    public class Bead
+    {
+        public Vector3 position;
+
+        public Bead(Vector3 position)
+        {
+            this.position = position;
+        }
+    }
+}
