+++ Assets/Scripts/Domain/LinkComponent.cs
@@ -1,0 +1,14 @@
+using System.Collections.Generic;
+
+namespace Domain
+{
+    public class LinkComponent
+    {
+        public readonly List<Bead> BeadList;
+
+        public LinkComponent(List<Bead> beadList)
+        {
+            BeadList = beadList;
+        }
+    }
+}
