--- Assets/Scripts/ILinkBeadsProvider.cs
@@ -1,7 +1,0 @@
-using System.Collections.Generic;
-using UnityEngine;
-
-public interface ILinkBeadsProvider
-{
-    List<Vector3[]> GetBeadsList();
-}
