+++ Assets/Scripts/PDCodeGeneration/PDCodeBead.cs
@@ -1,4 +1,5 @@
 using System;
+using Domain;
 using UnityEngine;
 
 namespace PDCodeGeneration
